requirements.txt
cobs==1.2.1
pyserial==3.5
pytest==8.4.1
flask==3.1.1